Fixing order of plugin during dist release + not running pact in dist release

+++ babun-core/plugins/git/install.sh
@@ -9,8 +9,12 @@ installed_version=$(echo "$1" || echo "0")
 # COMPATIBILITY FIXES
 # INSTALLED_VERSION=1
 if [[ "$installed_version" -le 1 ]]; then
-	echo "Compatibility fixes [git] version=[$installed_version]"
-	echo "Installing missing gettext and libsasl2"
-	pact install gettext || echo "Failed to install gettext"
-	pact install libsasl2 || echo "Failed to install libsasl2"
+	if [[ -f ~/.pact/pact.repo ]]; then
+		echo "Compatibility fixes [git] version=[$installed_version]"
+		echo "Installing missing gettext and libsasl2"	
+		pact install gettext || echo "Failed to install gettext"
+		pact install libsasl2 || echo "Failed to install libsasl2"
+	else 
+		echo "Compatibility fixes [git] version=[$installed_version] skipped. Not running in update mode!"
+	fi
 fi
